In case suspended before 120 and woke after 120

+++ scripts/cron/dine.sh
@@ -1,4 +1,8 @@
 #!/bin/dash
+time=$(date +%H%S)
+if [ "$time" -lt 2049 ] || [ "$time" -gt 2051 ] ; then
+    exit
+fi
 trap 'rm -f /tmp/sleep_dine.pid; exit' TERM
 echo "$$" >/tmp/sleep_dine.pid
 id=$(dunstify -p -t 0 "wrap up, it's time for dinner")
