+++ scripts/cron/sleep.sh
@@ -1,4 +1,8 @@
 #!/bin/dash
+time=$(date +%H%S)
+if [ "$time" -lt 0119 ] || [ "$time" -gt 0121 ] ; then
+    exit
+fi
 trap 'rm -f /tmp/sleep_dine.pid; exit' TERM
 echo "$$" >/tmp/sleep_dine.pid
 id=$(dunstify -p -t 0 "wrap up, it's time to go to bed")
